cli: modify get_distro_name to return the version id
Debmonitor is now being used to track hosts and Docker images,
and for the latter a number of questions/issues rose in the past
months:
- How many Docker images in the registry running a not supported
  OS like Stretch?
- How many Docker images in the registry should we cleanup/delete
  because running old and not-supported Debian distros?
There were a surprisingly high number of important services running
Stretch up to few months ago (Eventgate and Changeprop were the most
notable examples), so we should have an easier way to track
what we deploy in k8s-land.

This change updates get_distro_name to get_distro_version, so
that the OS column in the Debmonitor's report will change from
being simply "Debian" to something like "Debian 12".

Bug: T240193
Bug: T367427
Change-Id: If0e5d748e9b22f8595aedb7bef156e4a971e9d86