Do not use a login shell when dropping privileges
When running `apt install` a distracting warning message is displayed:

su: warning: cannot change directory to /nonexistent: No such file or
directory

We started having the issue with Debian Buster hosts and Moritz
Mühlenhoff found the cause for the new warning:

> In Debian releases up to Stretch, the su binary is built from the
> shadow source package. But starting with Buster it switched to the su
> implementation shipped in util-linux (which is totally different).

The reason is `su` is invoked with `-` which starts the shell as a login
shell and per documentation it "changes to the target user's home
directory". Given the `debmonitor` intentionally has no home directory,
that triggers the warning.

The `-` argument also:

* clears all the environment variables except TERM and variables
  specified by --whitelist-environment
* initializes the environment variables HOME, SHELL, USER, LOGNAME, and PATH
* changes to the target user’s home directory
* sets argv[0] of the shell to - in order to make the shell a login shell

The notable differences without `-` are:
* there is no change of directory, which addresses the warning
* the environment variables are NOT cleared (but `HOME`, `SHELL`,
  `USER` and `PATH` are still set). I don't think it would cause an
  issue.

`su -` was introduced in 2018 in operations/software/debmonitor by the
initial debianization: I4fbb6845d6ce1b349e131d0d2698ada87eed2b05

Bug: T216832
Change-Id: I7cf8b5fed331631a8eacbcb2d21fe019d2ed3466